+++ server.go
@@ -302,11 +302,16 @@ func (s *Server) load() error {
 			}
 		}
 
-		// Open all shards owned by server.
+		// Load shards.
+		s.shards = make(map[uint64]*Shard)
 		for _, db := range s.databases {
 			for _, rp := range db.policies {
 				for _, g := range rp.shardGroups {
 					for _, sh := range g.Shards {
+						// Add to lookups.
+						s.shards[sh.ID] = sh
+
+						// Only open shards owned by the server.
 						if !sh.HasDataNodeID(s.id) {
 							continue
 						}